requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
daphne==4.1.2
Django==5.0.6
django-autocomplete-light==3.11.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-listing==0.7.33
djangorestframework==3.15.1
dlib==19.24.4
drf-yasg==1.21.7
face-alignment==1.4.1
face-recognition==1.3.0
face_recognition_models==0.3.0
ffmpeg-python==0.2.0
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.23.3
hyperlink==21.0.0
idna==3.7
imageio==2.34.2
incremental==22.10.0
inflection==0.5.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.4.0
MouseInfo==0.1.3
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==2.0.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
piexif==1.1.3
pillow==10.3.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
rarfile==4.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rubicon-objc==0.4.9
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
selenium==4.22.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
sympy==1.12.1
tablib==3.6.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.24
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
trio==0.26.0
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
ultralytics==8.2.48
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==2.2.1
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
zope.interface==6.4.post2